Backchannel logout package refactored.
